+++ src/resources/components/textfield/mdc-text-field.js
@@ -1,6 +1,5 @@
 import { bindable, customElement, inject, DOM, bindingMode } from 'aurelia-framework';
 import { MDCTextField } from '@material/textfield/dist/mdc.textfield.min';
-import { MDCNotchedOutline } from '@material/notched-outline/dist/mdc.notchedOutline.min';
 
 @customElement('mdc-text-field')
 @inject(Element)
@@ -15,6 +14,8 @@ export class MdcTextField {
 	@bindable step;
 	@bindable min;
 	@bindable max;
+	@bindable leading;
+	@bindable trailing;
 
     constructor ( element) 
     {